+++ evm-contracts/src/v0.6/dev/AggregatorProxy.sol
@@ -84,6 +84,26 @@ contract AggregatorProxy is AggregatorInterface, Owned {
     return _latestRound();
   }
 
+  /**
+   * @notice get all details about a round. Consumers are encouraged to check
+   * that they're receiving fresh data by inspecting the updatedAt and
+   * answeredInRound return values.
+   * @param _roundId the round ID to retrieve the details for. If _roundId
+   * has the special value UINT256_MAX (2**256-1), the contract will retrieve
+   * the latest round's details.
+   * @return roundId is the round ID for which details were retrieved
+   * @return answer is the answer for the given round
+   * @return startedAt is the timestamp when the round was started. This is 0
+   * if the round hasn't been started yet.
+   * @return updatedAt is the timestamp when the round last was updated (i.e.
+   * answer was last computed)
+   * @return answeredInRound is the round ID of the round in which the answer
+   * was computed. answeredInRound may be smaller than roundId when the round
+   * timed out. answerInRound is equal to roundId when the round didn't time out
+   * and was completed regularly.
+   * @dev Note that for in-progress rounds (i.e. rounds that haven't yet received
+   * maxSubmissions) answer and updatedAt may change between queries.
+   */
   function getRoundData(uint256 _roundId)
     external
     view
@@ -92,8 +112,8 @@ contract AggregatorProxy is AggregatorInterface, Owned {
     returns (
       uint256 roundId,
       int256 answer,
-      uint64 startedAt,
-      uint64 updatedAt,
+      uint256 startedAt,
+      uint256 updatedAt,
       uint256 answeredInRound
     )
   {
@@ -123,8 +143,8 @@ contract AggregatorProxy is AggregatorInterface, Owned {
     returns (
       uint256 roundId,
       int256 answer,
-      uint64 startedAt,
-      uint64 updatedAt,
+      uint256 startedAt,
+      uint256 updatedAt,
       uint256 answeredInRound
     )
   {
@@ -204,8 +224,8 @@ contract AggregatorProxy is AggregatorInterface, Owned {
     returns (
       uint256 roundId,
       int256 answer,
-      uint64 startedAt,
-      uint64 updatedAt,
+      uint256 startedAt,
+      uint256 updatedAt,
       uint256 answeredInRound
     )
   {
@@ -218,8 +238,8 @@ contract AggregatorProxy is AggregatorInterface, Owned {
     returns (
       uint256 roundId,
       int256 answer,
-      uint64 startedAt,
-      uint64 updatedAt,
+      uint256 startedAt,
+      uint256 updatedAt,
       uint256 answeredInRound
     )
   {
