+++ evm-contracts/src/v0.6/dev/WhitelistedAggregator.sol
@@ -110,8 +110,8 @@ contract WhitelistedAggregator is FluxAggregator, Whitelisted {
     returns (
       uint256 roundId,
       int256 answer,
-      uint64 startedAt,
-      uint64 updatedAt,
+      uint256 startedAt,
+      uint256 updatedAt,
       uint256 answeredInRound
     )
   {
@@ -145,8 +145,8 @@ contract WhitelistedAggregator is FluxAggregator, Whitelisted {
     returns (
       uint256 roundId,
       int256 answer,
-      uint64 startedAt,
-      uint64 updatedAt,
+      uint256 startedAt,
+      uint256 updatedAt,
       uint256 answeredInRound
     )
   {
