+++ evm-contracts/src/v0.6/dev/WhitelistedFacade.sol
@@ -97,8 +97,8 @@ contract WhitelistedFacade is AggregatorFacade, Whitelisted {
     returns (
       uint256 roundId,
       int256 answer,
-      uint64 startedAt,
-      uint64 updatedAt,
+      uint256 startedAt,
+      uint256 updatedAt,
       uint256 answeredInRound
     )
   {
@@ -129,8 +129,8 @@ contract WhitelistedFacade is AggregatorFacade, Whitelisted {
     returns (
       uint256 roundId,
       int256 answer,
-      uint64 startedAt,
-      uint64 updatedAt,
+      uint256 startedAt,
+      uint256 updatedAt,
       uint256 answeredInRound
     )
   {
